test(ValidatorTest.java): 유효성 체크 테스트
숫자야구게임 기능 중 유효성 체크 기능을 테스트 한다.
- validateStringNumbers() 에서 1~9 로 이루어진 서로 다른 3자리수 유효성 체크 테스트
- validateStrikeAndBallCounts() 에서 스트라이크, 볼 개수 유효성 체크 테스트
- validateRestartOrEndStatus() 에서 사용자 입력값에 따른 게임 재시작 또는 종료 상태 유효성 체크 테스트
